Tries to add Portainer Agent as HassIO Addon - not possible
As it is impossible for a HassIO Addon to mount "/" or docker.sock into the Addon container this the agent can't be provided as a Addon.

The idea was to have a "clean system" from HassIO perspective and do not have to run the agent "from the outside" as container on the host NOT managed by HassiO / Supervisor.

See the patch_supervisor.sh script for information about the possible "patch" to exclude some host containers from the container check. But also why we can just abdone the agent and therefore do not have to edit the main (HassIO) system!